requirements/llama.txt
#!/usr/bin/env -S pip install -r

jinja2
llama-cpp-python
